Move KeyResolver to it's own file
