+++ lib/feature_flagger.rb
@@ -8,6 +8,7 @@ require 'feature_flagger/control'
 require 'feature_flagger/model'
 require 'feature_flagger/model_settings'
 require 'feature_flagger/feature'
+require 'feature_flagger/key_resolver'
 require 'feature_flagger/configuration'
 require 'feature_flagger/manager'
 require 'feature_flagger/feature_keys_migration'
